# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Workyy Realtime API
  version: 0.2.0
servers:
  - url: https://api.workyy.dev
    description: Production
  - url: http://localhost:4000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Problem:
      type: object
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          example: https://api.workyy.dev/problems/conflict
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties: true
    Board:
      type: object
      required: [id, workspaceId, title, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Node:
      type: object
      required: [id, boardId, type, position]
      properties:
        id:
          type: string
          format: uuid
        boardId:
          type: string
          format: uuid
        type:
          type: string
          enum: [sql, python, table, plot]
        position:
          type: object
          required: [x, y]
          properties:
            x: { type: number }
            y: { type: number }
        payload:
          type: object
          nullable: true
          additionalProperties: true
    Edge:
      type: object
      required: [id, sourceId, targetId]
      properties:
        id:
          type: string
          format: uuid
        sourceId:
          type: string
          format: uuid
        targetId:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: true
...
paths:
  /boards:
    post:
      summary: Create a new board
      operationId: createBoard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, title]
              properties:
                workspaceId:
                  type: string
                  format: uuid
                title:
                  type: string
                description:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Board created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: Workspace not found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Board title conflict
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /boards/{boardId}:
    get:
      summary: Get board with nodes and edges
      operationId: getBoard
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: boardId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Board details
          content:
            application/json:
              schema:
                type: object
                required: [board, nodes, edges]
                properties:
                  board:
                    $ref: '#/components/schemas/Board'
                  nodes:
                    type: array
                    items: { $ref: '#/components/schemas/Node' }
                  edges:
                    type: array
                    items: { $ref: '#/components/schemas/Edge' }
        '404':
          description: Not found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...